Move screenshots handling to the separate class
